requirements.txt
python-dotenv==1.0.0
playwright==1.40.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.104.0
openai==1.3.7
requests==2.31.0
pyperclip==1.8.2
